mqtpro/requirements.txt
asgiref==3.6.0
channels==4.0.0
Django==4.2
django-mqtt==0.4.0
gevent==22.10.2
greenlet==2.0.2
paho-mqtt==1.6.1
sqlparse==0.4.3
websocket==0.2.1
websocket-client==1.5.1
zope.event==4.6
zope.interface==6.0
